requirements.txt
python-telegram-bot==20.0
matplotlib==3.7.1
python-dotenv==1.0.0
